auth fix logic
