+++ package.json
@@ -10,6 +10,7 @@
     "ngrok": "ngrok http http://localhost:3000"
   },
   "dependencies": {
+    "@auth/mongodb-adapter": "^2.3.3",
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.15.7",
@@ -22,6 +23,7 @@
     "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^4.0.0",
     "lodash": "^4.17.21",
+    "mongodb": "^6.3.0",
     "mongoose": "^8.1.1",
     "next": "14.1.0",
     "next-auth": "^4.24.5",
