+++ src/types/mongodb.d.ts
@@ -1,0 +1,5 @@
+import { MongoClient } from "mongodb";
+
+declare global {
+  var _mongoClientPromise: Promise<MongoClient>;
+}
